Fix to allow string messages to be passed and not only objects to produce
Object.assign({}, message) is used internally to transform the message,
but if the message is a string this will mangle it and turn into an object (Eg for "test" it wil become {0: "t", 1: "e", 2: "s", 3: "t"}) and this was causing problems in some of our services.
Fix is to check if message is a string and not process with Object.assign if it is. Also added test.